outbound: Fix incorrect l5d-proxy-connection logs
We've received reports of proxies inexplicibly emiting log lines
including "Received unmeshed response with l5d-proxy-connection set".
These messages may arise when the endpoint stack returns a synthesized
response.

Furthermore, we have a note in the code explaining that this connection
closure logic should not apply to load balanced requests, though it
currently is.

This change fixes both of these issues:

* The proxy_connection_close module is renamed to
  handle_proxy_error_headers.
* It is now only applied in the endpoint stack. It doesn't make any
  sense for it to be applied in the server stack, since we'll already
  have cleared any headers set by peers. Removing this module prevents
  the application of teardown logic on synthetic responses.
* `NewHandleProxyErrorHeaders` is now configured by a `Closable`
  parameter that determines whether teardown logic should be applied.
  This parameter is only enabled when forwarding to a single endpoint.
  No teardown logic is applied when load balancing.
* In a future change, we should stop emitting l5d-proxy-connection when
  synthesizing outbound responses.
