+++ linkerd/app/outbound/src/http/concrete.rs
@@ -1,7 +1,7 @@
 //! A stack that (optionally) resolves a service to a set of endpoint replicas
 //! and distributes HTTP requests among them.
 
-use super::{balance, client};
+use super::{balance, client, handle_proxy_error_headers};
 use crate::{http, stack_labels, Outbound};
 use linkerd_app_core::{
     metrics, profiles,
@@ -41,6 +41,7 @@ pub struct Endpoint<T> {
     is_local: bool,
     metadata: Metadata,
     parent: T,
+    closable: bool,
 }
 
 /// A target configuring a load balancer stack.
@@ -129,6 +130,10 @@ impl<N> Outbound<N> {
                             metadata,
                             is_local,
                             parent: target.parent,
+                            // We don't close server-side connections when we
+                            // get `l5d-proxy-connect: close` response headers
+                            // going through the balancer.
+                            closable: false,
                         }
                     }
                 })
@@ -158,6 +163,7 @@ impl<N> Outbound<N> {
                                     addr,
                                     metadata,
                                     parent,
+                                    closable: true,
                                 }
                             }),
                         })
@@ -217,6 +223,12 @@ impl<T> svc::Param<Option<http::AuthorityOverride>> for Endpoint<T> {
             .authority_override()
             .cloned()
             .map(http::AuthorityOverride)
+    }
+}
+
+impl<T> svc::Param<handle_proxy_error_headers::Closable> for Endpoint<T> {
+    fn param(&self) -> handle_proxy_error_headers::Closable {
+        handle_proxy_error_headers::Closable(self.closable)
     }
 }
